Align .Net BatchWriteItem Class name with the file
